ID-8 Regroup unit tests
